% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{accrossAssessorReport}
\alias{accrossAssessorReport}
\title{accrossAssessorReport}
\usage{
accrossAssessorReport(semAllAssessordf, nAssesor, nParticipant)
}
\arguments{
\item{semAllAssessordf}{[dataframe] standard error of measurement table computed for ALL assessors}

\item{nAssesor}{[integer] number of assessors}

\item{nParticipant}{[integer] number of participants}
}
\value{
[]
}
\description{
Report of across asessor differences
}
\section{Warning}{

}

\examples{

}
