% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{addCumulatedCycleindex}
\alias{addCumulatedCycleindex}
\title{addCumulatedCycleindex}
\usage{
addCumulatedCycleindex(table, groupByList)
}
\arguments{
\item{table}{[dataframe] all-cycle table.}

\item{groupByList}{[list] group-by independant variables .}
}
\value{
update all-cycle table
}
\description{
add new col CycleCum, cumulating cycle number from both left and right contexts
}
\examples{
addCumulatedCycleindex(kinematicTable, c("Id"))
}
