% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{addGaitDescriptiveEventsLines}
\alias{addGaitDescriptiveEventsLines}
\title{addGaitDescriptiveEventsLines}
\usage{
addGaitDescriptiveEventsLines(fig, descStatsPhaseTable, iContext,
  colorFactor = NULL, linetypeFactor = NULL)
}
\arguments{
\item{colorFactor}{[string] line color according an independant variable}

\item{linetypeFactor}{[string] line type definied according an independant variable}

\item{[fig]}{ggplot2 figure}

\item{[descStatsPhaseTable]}{descriptive stats table of gait phase scalar
(must include c("stancePhase","doubleStance1","doubleStance2")}

\item{[Context]}{selected context}
}
\value{
[]
}
\description{
add gait events of a specific context as vertical lines to a figure
}
\section{Warning}{

}

\examples{

}
