% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{betweenAssessors}
\alias{betweenAssessors}
\title{betweenAssessors}
\usage{
betweenAssessors(sessionAverage)
}
\arguments{
\item{sessionAverage}{[dataframe] session average table}
}
\description{
assess between assessor difference
}
\examples{


}
