% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{betweenAssessorsReport}
\alias{betweenAssessorsReport}
\title{betweenAssessorsReport}
\usage{
betweenAssessorsReport(betweenAssessorAssement)
}
\arguments{
\item{betweenAssessorAssement}{[dataframe] between assessor assessment table}
}
\description{
report of  between assessor differences
}
\examples{



}
