% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descritiveStats_LowLevel.R
\name{computeDescritiveStats}
\alias{computeDescritiveStats}
\title{computeDescritiveStats}
\usage{
computeDescritiveStats(table, dependantVariables, groupByList,
  bySubjectFlag = TRUE)
}
\arguments{
\item{table}{[dataframe] all-cycle table.}

\item{dependantVariables}{[list] dependant variables}

\item{groupByList}{[list] group-by independant variables}

\item{bySubjectFlag}{[bool] group results by subject}
}
\value{
table
}
\description{
return descriptive statistics for a series of dependant variables and a combinaison of factors
}
\examples{
TODO
}
