% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descritiveStats_LowLevel.R
\name{computeDescritiveStats_onFrameSequences}
\alias{computeDescritiveStats_onFrameSequences}
\title{computeDescritiveStats_onFrameSequences}
\usage{
computeDescritiveStats_onFrameSequences(table, groupByList,
  bySubjectFlag = TRUE)
}
\arguments{
\item{table}{[dataframe] all-cycle table}

\item{groupByList}{[list] group-by independant variables}

\item{bySubjectFlag}{[bool] output table grouped by subject}
}
\value{
table
}
\description{
return descriptive stats for all frame sequences
}
\examples{
TODO
}
