% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{computeMetrics_onFrameSequence}
\alias{computeMetrics_onFrameSequence}
\title{metrics_onFrames_local}
\usage{
computeMetrics_onFrameSequence(table, metricsFunction, Label, Context,
  Axes, modality1, modality2, comparisonLabel = NULL)
}
\arguments{
\item{table}{[dataframe] all-cycle table}

\item{metricsFunction}{[string] metrics ( eg, mae, rmse) to apply}

\item{Label}{[string] label of the frame sequence}

\item{Context}{[string] context of the frame sequence}

\item{Axes}{[string] axis of the frame sequence}

\item{modality1}{[string] 1st selected modality of the independant variable (ComparisonFactor)}

\item{modality2}{[string] 2nd selected modality of the independant variable (ComparisonFactor)}

\item{comparisonLabel}{[string] label given to the comparison}

\item{frameIndexes}{[list] selected frame indexes}
}
\value{
[dataframe]
}
\description{
compute basic metrics between two modalities of the ComparisonFactor Factor for a specific point
}
\section{Warning}{

need construction of the factor ComparisonFactor
}

\examples{
computeMetrics_onFrameSequence(kinematicTable,"rmse", "LKneeAngles","Left",c("X","Y","Z"),"Rigid","THIsta")
}
