% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{computeMetrics_onFrameSequences}
\alias{computeMetrics_onFrameSequences}
\title{computeMetrics_onFrameSequences}
\usage{
computeMetrics_onFrameSequences(table, metricsFunction, modality1,
  modality2, comparisonLabel = NULL, frameIndexes = NULL)
}
\arguments{
\item{table}{[dataframe] all-cycle table}

\item{metricsFunction}{[string] metrics ( eg, mae, rmse) to apply}

\item{modality1}{[string] 1st selected modality of the independant variable (ComparisonFactor)}

\item{modality2}{[string] 2nd selected modality of the independant variable (ComparisonFactor)}

\item{comparisonLabel}{[string] label given to the comparison}

\item{frameIndexes}{[list] selected frame indexes}
}
\value{
[dataframe]
}
\description{
compute basic metrics between two modalities of the independant variable (\textbf{ComparisonFactor})
 for all frame sequences
}
\section{Warning}{

need construction of the factor ComparisonFactor
}

\examples{
computeMetrics_onFrameSequences(kinematicTable,"rmse",  "modality1","modality2",comparisonLabel = "mod1_mod2")
}
