% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{computeMetrics_onScalar}
\alias{computeMetrics_onScalar}
\title{computeMetrics_onScalar}
\usage{
computeMetrics_onScalar(table, metricsFunction, DiscreteLabel, modality1,
  modality2, comparisonLabel = NULL)
}
\arguments{
\item{table}{[dataframe] all-cycle table}

\item{metricsFunction}{[string] metrics ( eg, mae, rmse) to apply}

\item{modality1}{[string] 1st selected modality of the independant variable (ComparisonFactor)}

\item{modality2}{[string] 2nd selected modality of the independant variable (ComparisonFactor)}

\item{comparisonLabel}{[string] label given to the comparison}
}
\value{
[dataframe]
}
\description{
compute basic metrics between two modalities of the independant variable (\textbf{ComparisonFactor})
 for a specific Scalar
}
\section{Warning}{

need construction of the factor ComparisonFactor


need construction of the factor ComparisonFactor
}

\examples{
TODO
}
