% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{consistencyPlot}
\alias{consistencyPlot}
\title{consistencyPlot}
\usage{
consistencyPlot(table, iContext, iLabel, iAxis, iTitle = "",
  yLabel = "Deg", legendPosition = "none", ylimits = NULL,
  colorFactor = NULL, facetFactor = NULL, linetypeFactor = NULL)
}
\arguments{
\item{table}{[dataframe] all-cycle table}

\item{iContext}{[string] context of the frame sequence}

\item{iLabel}{[string] label of the frame sequence}

\item{iAxis}{[string] axis of the frame sequence}

\item{iTitle}{[string] plot title}

\item{yLabel}{[string] label of the Y-axis}

\item{legendPosition}{[string] position of the legend (see legend.position of ggplot2)}

\item{ylimits}{[list] limits of the y-axis}

\item{colorFactor}{[string] line color according an independant variable}

\item{facetFactor}{[string] create \textit{facet} plot ( see ggplot2) for an  independant variable}

\item{linetypeFactor}{[string] line type definied according an independant variable}
}
\value{
fig [ggplot2 figure]
}
\description{
plot all cycle of a frame sequence
}
\section{Warning}{

}

\examples{

}
