% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{constructOverallTable}
\alias{constructOverallTable}
\title{constructOverallTable}
\usage{
constructOverallTable(table)
}
\arguments{
\item{table}{[dataframe] all-cycle table.}
}
\value{
update all-cycle table.
}
\description{
merge left and right contexts. \code{Left} and \code{Right} modalities are replaced by \code{Overall}
}
\section{Warning}{

function addCumulatedCycleindex needs to be run before
}

\examples{
constructOverallTable(kinematicTable)
}
