% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{constructTableFromXls}
\alias{constructTableFromXls}
\title{constructTableFromXls}
\usage{
constructTableFromXls(fullXlsFiles, sheet)
}
\arguments{
\item{sheet}{[String] sheet name}

\item{fullXlsFile}{[String] full filename (path+fileame) of excel spreadsheets}
}
\value{
[]
}
\description{
load annd concatenate a list of excel spreadsheet
}
\examples{


}
