% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descritiveStats_HighLevel.R
\name{descriptiveGait}
\alias{descriptiveGait}
\title{descriptiveGait}
\usage{
descriptiveGait(table, groupByList, bySubjectFlag = TRUE)
}
\arguments{
\item{table}{[dataframe] all-cycle table}

\item{bySubjectFlag}{[bool] group-by independant variables}
}
\value{
[list] Frames : descriptive stats of frame-based variables , Events : descriptive stats of events
}
\description{
return descriptive stats of both frame sequence and gait event scalars
}
\examples{
TODO

}
