% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPanel.R
\name{descriptiveKinematicGaitPanel}
\alias{descriptiveKinematicGaitPanel}
\title{descriptiveKinematicGaitPanel}
\usage{
descriptiveKinematicGaitPanel(descStatsFrameSequence, descStatsPhases,
  iContext, colorFactor = NULL, linetypeFactor = NULL,
  normativeData = NULL, stdCorridorFlag = FALSE,
  manualLineType = NULL, manualSizeType = NULL)
}
\arguments{
\item{descStatsFrameSequence}{[dataframe] descriptive stats table of all frame sequences}

\item{descStatsPhases}{[dataframe] descriptive stats table of gait phase scalar ()}

\item{iContext}{[string] context of the frame sequence}

\item{colorFactor}{[string] line color according an independant variable}

\item{linetypeFactor}{[string] line type definied according an independant variable}

\item{normativeData}{[dataframe] table of a normative dataset}

\item{stdCorridorFlag}{[Bool] add std corridor to plot}

\item{manualLineType}{[list] manual line type ( see ggplot2 doc)}

\item{manualSizeType}{[float] manual line size ( see ggplot2 doc)}
}
\value{
fig [ggplot2 figure]
}
\description{
convenient descriptive plot panel of gait kinematics for a specific context
}
\section{Warning}{

}

\examples{

}
