% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPanel.R
\name{descriptiveKineticGaitPanel_bothContext}
\alias{descriptiveKineticGaitPanel_bothContext}
\title{descriptiveKineticGaitPanel_bothContext}
\usage{
descriptiveKineticGaitPanel_bothContext(descStatsFrameSequence,
  descStatsPhases = NULL, normativeData = NULL,
  stdCorridorFlag = FALSE)
}
\arguments{
\item{descStatsFrameSequence}{[dataframe] descriptive stats table of all frame sequences}

\item{descStatsPhases}{[dataframe] descriptive stats table of gait phase scalar ()}

\item{normativeData}{[dataframe] table of a normative dataset}

\item{stdCorridorFlag}{[Bool] add std corridor to plot}
}
\value{
fig [ggplot2 figure]
}
\description{
convenient descriptive plot panel of gait kinematics for left and right contexts
}
\section{Warning}{

}

\examples{

}
