% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{descriptivePlot}
\alias{descriptivePlot}
\title{descriptivePlot}
\usage{
descriptivePlot(descStatsFrameSequence, iContext, iLabel, iAxis,
  iTitle = "", yLabel = "Deg", legendPosition = "none",
  colorFactor = NULL, facetFactor = NULL, linetypeFactor = NULL,
  ylimits = NULL, lineWidth = 0.5)
}
\arguments{
\item{descStatsFrameSequence}{[dataframe] descriptive stats table of frame sequence}

\item{iContext}{[string] context of the frame sequence}

\item{iLabel}{[string] label of the frame sequence}

\item{iAxis}{[string] axis of the frame sequence}

\item{iTitle}{[string] plot title}

\item{yLabel}{[string] label of the Y-axis}

\item{legendPosition}{[string] position of the legend (see legend.position of ggplot2)}

\item{colorFactor}{[string] line color according an independant variable}

\item{facetFactor}{[string] create \textit{facet} plot ( see ggplot2) for an  independant variable}

\item{linetypeFactor}{[string] line type definied according an independant variable}

\item{ylimits}{[list] limits of the y-axis}

\item{lineWidth}{[float] line width value}
}
\value{
fig [ggplot2 figure]
}
\description{
plot descriptive statistics of a frame sequence
}
\section{Warning}{

}

\examples{

}
