% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{descriptivePlot_bothContext}
\alias{descriptivePlot_bothContext}
\title{descriptivePlot_bothContext}
\usage{
descriptivePlot_bothContext(descStatsFrameSequence, LabelLeft, AxisLeft,
  LabelRight, AxisRight, iTitle = "", yLabel = "Deg",
  legendPosition = "none", ylimits = NULL)
}
\arguments{
\item{descStatsFrameSequence}{[dataframe] descriptive stats table of frame sequence}

\item{LabelLeft}{[string] label of the frame sequence for the left context}

\item{AxisLeft}{[string] axis of the frame sequence for the left context}

\item{LabelRight}{[string] label of the frame sequence for the right context}

\item{AxisRight}{[string] axis of the frame sequence for the Left context}

\item{iTitle}{[string] plot title}

\item{yLabel}{[string] label of the Y-axis}

\item{legendPosition}{[string] position of the legend (see legend.position of ggplot2)}

\item{ylimits}{[list] limits of the y-axis}
}
\value{
fig [ggplot2 figure]
}
\description{
plot left and right descriptivestats  of a frame sequence
}
\section{Warning}{

}

\examples{

}
