% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descritiveStats_LowLevel.R
\name{gather_descriptiveStats_FrameSequences}
\alias{gather_descriptiveStats_FrameSequences}
\title{gather_descriptiveStats_FrameSequences}
\usage{
gather_descriptiveStats_FrameSequences(frameSeqDescStats)
}
\arguments{
\item{frameSeqDescStats}{[dataframe] frame sequence descriptive stats}
}
\value{
dataframe.
}
\description{
#' applicatio of dplyr::gather function  on frame sequence descriptive stats table.
( return a \textit{long} table)
}
\examples{
TODO
}
