% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descritiveStats_LowLevel.R
\name{gather_descritiveStats}
\alias{gather_descritiveStats}
\title{gather_DescritiveStats}
\usage{
gather_descritiveStats(descStatTable, dependantVariables, groupByList)
}
\arguments{
\item{descStatTable}{[dataframe] descriptive stats table.}

\item{dependantVariables}{[list] dependant variables}

\item{groupByList}{[list] group-by independant variables}
}
\value{
table
}
\description{
applicatio of dplyr::gather function  on descriptive stats table. ( return a \textit{long} table)
}
\examples{
TODO
}
