% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{geom_normative_ribbon}
\alias{geom_normative_ribbon}
\title{geom_normative_ribbon}
\usage{
geom_normative_ribbon(data)
}
\arguments{
\item{data}{[dataframe] normative data table}
}
\value{
[]
}
\description{
new ggplot2 geom displaying  corridor of normative data
}
\section{Note}{

programming as a new geom ( see https://rpubs.com/hadley/97970)
}

\examples{

}
