% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{geom_stdRibbon}
\alias{geom_stdRibbon}
\title{geom_stdRibbon}
\usage{
geom_stdRibbon(table)
}
\arguments{
\item{table}{[dataframe] all cycle table}
}
\value{
[]
}
\description{
new ggplot2 geom displaying std corridor
}
\section{Note}{

programming as a new geom ( see https://rpubs.com/hadley/97970)
}

\examples{

}
