% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{geom_vline_descriptiveEvents_bothContext}
\alias{geom_vline_descriptiveEvents_bothContext}
\title{addGaitDescriptiveEventsLines_bothContext}
\usage{
geom_vline_descriptiveEvents_bothContext(descStatsPhaseTable)
}
\arguments{
\item{[fig]}{ggplot2 figure}

\item{[descStatsPhaseTable]}{descriptive stats table of gait phases
(must include c("stancePhase","doubleStance1","doubleStance2")}
}
\value{
[]
}
\description{
add gait events from left and right context as vertical lines to a figure
}
\section{Warning}{

}

\examples{

}
