% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varianceEstimation.R
\name{getGlobalVariance}
\alias{getGlobalVariance}
\title{getGlobalVariance}
\usage{
getGlobalVariance(varianceTable)
}
\arguments{
\item{[]}{}
}
\value{
[]
}
\description{

}
\section{Warning}{

}

\examples{

}
