% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{getNumberOfCycles}
\alias{getNumberOfCycles}
\title{getNumberOfCycles}
\usage{
getNumberOfCycles(table, label, axis, groupByList)
}
\arguments{
\item{table}{[dataframe] all-cycle table}

\item{label}{[string] label of the frame sequence}

\item{axis}{[string] axis of the frame sequence}

\item{groupByList}{[list] group-by independant variables}
}
\value{
dataframe.
}
\description{
return number of cycles for a label, its axis and a combinaison of factor
}
\examples{
getNumberOfCycles(kinematicTable,"LHipAngles","X", c("Id","Operator","Session"))
}
