% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descritiveStats_LowLevel.R
\name{getStdCorridorLimits_fromDescStatFrameSequences}
\alias{getStdCorridorLimits_fromDescStatFrameSequences}
\title{getStdCorridorLimits_fromDescStatFrameSequences}
\usage{
getStdCorridorLimits_fromDescStatFrameSequences(frameSeqDescStats)
}
\arguments{
\item{frameSeqDescStats}{[dataframe] frame sequence descriptive stats}
}
\value{
dataframe.
}
\description{
return corridor limits of frame seqence descriptive stats
}
\examples{
TODO
}
