% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{homogeniseCycles}
\alias{homogeniseCycles}
\title{homogeniseCycles}
\usage{
homogeniseCycles(table, groupByList)
}
\arguments{
\item{table}{[dataframe] all-cycle table}

\item{groupByList}{[list] group-by independant variables .}
}
\value{
update all-cycle table
}
\description{
homogenise the number of cycle for a combinaison of independant variables
}
\examples{
homogeniseCycles(kinematicTable, c("Id","Operator","Session"))
}
