% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{LinearFit_onFrameSequences}
\alias{LinearFit_onFrameSequences}
\title{LinearFit_onFrameSequences}
\usage{
LinearFit_onFrameSequences(table, modality1, modality2,
  comparisonLabel = NULL, frameIndexes = NULL)
}
\arguments{
\item{table}{[dataframe] all-cycle table}

\item{modality1}{[string] 1st selected modality of the independant variable (ComparisonFactor)}

\item{modality2}{[string] 2nd selected modality of the independant variable (ComparisonFactor)}

\item{comparisonLabel}{[string] label given to the comparison}

\item{frameIndexes}{[list] selected frame indexes}
}
\value{
[dataframe]
}
\description{
Linear fitting between two modalities of the independant variable (\textbf{ComparisonFactor})
}
\section{Warning}{

need construction of the factor ComparisonFactor
}

\examples{
TODO
}
