% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{loadNormativeDataSet}
\alias{loadNormativeDataSet}
\title{loadNormativeDataSet}
\usage{
loadNormativeDataSet(fullXlsFile, sheet)
}
\arguments{
\item{fullXlsFile}{[String] full filename (path+fileame) of the selected dataset}
}
\value{
[]
}
\description{
load  normative dataset
}
\examples{


}
