% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{MeanAbsoluteVariability}
\alias{MeanAbsoluteVariability}
\title{MeanAbsoluteVariability}
\usage{
MeanAbsoluteVariability(table)
}
\arguments{
\item{table}{[dataframe] all-cycle table}
}
\value{
[dataframe]
}
\description{
compute the mean absolute variability (see Mantovani2016) for all modalities
of the independant variables  (\textbf{ComparisonFactor})
}
\section{Warning}{

need construction of the factor ComparisonFactor
}

\examples{
MeanAbsoluteVariability(kinematicCyles)
}
