% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{MinDetectableChange}
\alias{MinDetectableChange}
\title{MinDetectableChange f}
\usage{
MinDetectableChange(table, dependantVariables)
}
\arguments{
\item{table}{[dataframe] all-cycle table}

\item{dependantVariables}{[list] dependant variables}
}
\value{
[dataframe]
}
\description{
compute the min detectable change (see Mantovani2016) for specific dependant variables
}
\section{Warning}{

need construction of the factor ComparisonFactor
}

\examples{
mdc = MinDetectableChange(kinematicTable,c("Frame0","Frame001") )
}
