% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normativeDataSet_transformation.R
\name{normativeDataSetTransformation}
\alias{normativeDataSetTransformation}
\title{normativeDataSetTransformation}
\usage{
normativeDataSetTransformation(normativeTable)
}
\arguments{
\item{normativeTable}{[dataframe] normative table}
}
\value{
dataframe.
}
\description{
transform  the normative table  as long table with new columns max (mean+std) and min (mean-std)
}
\examples{
TODO
}
