% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{Sem_allAssessors}
\alias{Sem_allAssessors}
\title{Sem_allAssessors}
\usage{
Sem_allAssessors(sessionAverage, Anova = TRUE)
}
\arguments{
\item{sessionAverage}{[dataframe] session average table}

\item{Anova}{[Bool] computation done with anova}
}
\description{
compute the standard error of measurement  for all assessors
}
\examples{

}
