% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{Sem_byAssessor}
\alias{Sem_byAssessor}
\title{Sem_byAssessor}
\usage{
Sem_byAssessor(sessionAverage, Anova = TRUE)
}
\arguments{
\item{sessionAverage}{[dataframe] session average table}

\item{Anova}{[Bool] computation done with anova}
}
\description{
compute the standard error of meausrement by assessors
}
\examples{

}
