% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{withinAssessorReport}
\alias{withinAssessorReport}
\title{withinAssessorReport}
\usage{
withinAssessorReport(sembyAssessordf, nAssesor, nParticipant)
}
\arguments{
\item{sembyAssessordf}{[dataframe] standard error of measurement table computed by assessor}

\item{nAssesor}{[integer] number of assessors}

\item{nParticipant}{[integer] number of participants}
}
\value{
[]
}
\description{
Report of the within asessor differences
}
\section{Warning}{

}

\examples{

}
