% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{withinSubjectStandardDeviation}
\alias{withinSubjectStandardDeviation}
\title{withinSubjectStandardDeviation}
\usage{
withinSubjectStandardDeviation(sessionAverage, Anova = FALSE)
}
\arguments{
\item{sessionAverage}{[dataframe] session average table}

\item{Anova}{[Bool] computation done with anova}
}
\value{
[dataframe]
}
\description{
compute the within-subject standard deviation
}
\section{Warning}{

within-subject standard deviation matches Richard Baker's spreadsheet processing.
You might also compute this value through  Anova outputs. In this case, slight differences might occur
}

\examples{

}
